Correlation of Interaction Similarity for connector species

Calculating correlation of interaction similarity for connector species.

cois(
  network.or.subnet_mat1,
  subnet_mat2 = NULL,
  weighted = FALSE,
  method = "kendall"
)

Arguments

network.or.subnet_mat1

Igraph object or matrix. An "igraph" object with node attribute 'level' or a matrix representing one subnetwork.

subnet_mat2

The matrix representing one subnetwork.

weighted

Logical. Default to FALSE. See details.

method

Correlation method ("pearson", "kendall" or "spearman"). Default to "kendall".

Value

Return a numeric value representing correlation of interaction similarity for connector species.

Details

In this package, a tripartite network contains three groups of nodes (a-nodes,b-nodes,c-nodes) and two subnetworks (P includes the links between a-nodes and b-nodes, Q includes the links between b-nodes and c-nodes). Connector nodes belong to b-nodes.

This function follows Sauve et al.(2016) to calculate the correlation of interaction similarity of connector nodes. When weighted=FALSE, the Jaccard similarity coefficients between all pairs of connector nodes in each subnetwork, defined as the number of shared interaction partners divided by the total number of interaction partners involved (i.e. the maximum number of interactions they could share). When weighted=TRUE, the quantitative similarities of all pairs of connector nodes in each subnetwork were measured using the generalised Jaccard coefficient. Three correlation methods are provided. Kendall correlation is recommended following Sauve et al.(2016).

If the weighted = FALSE, the input network can be an "igraph" object or two matrices. If the weighted = TRUE, only matrices are accepted.

Two types of inputs network.or.subnet_mat1 can be processed:

An "igraph" object with node attribute 'level' (0 for a-nodes, 1 for b-nodes,2 for c-nodes)

Or a matrix representing subnetwork P, and must be input with subnet_mat2 representing subnetwork Q.

If the inputs are two matrices, please make sure the rows of network.or.subnet_mat1 and subnet_mat2 both represent the groups of connector species,i.e, the b-group species. If both matrices have row names, then the function matches row names to produce connector nodes. Otherwise, row numbers are assigned to row names and matched. Within the two matrices (P and Q), columns represents a-group nodes and c-group nodes respectively. Elements in matrices are non-zero values if two nodes are linked with or without weights, and 0 otherwise.

References

Sauve, A. M., ThC)bault, E., Pocock, M. J., & Fontaine, C. (2016). How plants connect pollination and herbivory networks and their contribution to community stability. Ecology, 97(4), 908-917.

Examples

## generate a random binary tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)
cois(Net)
#> CoIS= -0.0120350465896841;
#> [1] -0.01203505
## empirical network
data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
cois(Net)
#> CoIS= 0.475167309613609;
#> [1] 0.4751673
set.seed(13)
E(Net)$weight<-runif(length(E(Net)),0.1,1)#random weights assigned
#> Error in E(Net): could not find function "E"
cois(Net,weighted=T)
#> CoIS_weighted= 0.475167309613609;
#> [1] 0.4751673

##input as binary matrices,with row names.
md1 <- matrix(sample(c(0,1),8*11,replace=TRUE),8,11,dimnames = list(paste0("b",1:8),paste0("c",1:11)))
md2 <- matrix(sample(c(0,1),10*12,replace=TRUE),10,12,dimnames = list(paste0("b",1:10),paste0("a",1:12)))
cois(md1,md2)
#> CoIS= 0.0650091129746962;
#> [1] 0.06500911

##input as weighted matrices,with row numbers as row names.
mdw1 <- matrix(sample(c(rep(0,40),runif(48,0,1))),8,11)
mdw2 <- matrix(sample(c(rep(0,40),runif(80,0,1))),10,12)
cois(mdw1,mdw2)
#> No rownames for matrices, so row IDs are used!
#> CoIS= 0.0168099544404158;
#> [1] 0.01680995
cois(mdw1,mdw2,weighted=T)
#> No rownames for matrices, so row IDs are used!
#> CoIS_weighted= 0.0264550264550265;
#> [1] 0.02645503