Counts of interconnection motifs

Counting the frequencies of interconnection motifs for a tripartite interaction network.

icmotif_count(network.or.subnet_mat1, subnet_mat2 = NULL, weighted = FALSE)

Arguments

network.or.subnet_mat1

Igraph object or matrix. An "igraph" object with node attribute 'level' or a matrix representing one subnetwork.

subnet_mat2

The matrix representing one subnetwork.

weighted

Logical. Default to FALSE. If TRUE, the counts of motifs are weighted by link strengths in each motif.

Value

return a numeric vector of the number of 48 interconnection motifs. The motifs are named ???MABC-i???: M means ???motif???, ???A??? is the number of a-nodes, ???B??? is the number of b-nodes, ???C??? is the number of c-nodes and ???i??? is the serial number for the motifs with the same ???ABC???. See 'Multi_motif' for the forms.

Details

In this package, a tripartite network contains three groups of nodes (a-nodes,b-nodes,c-nodes) and two subnetworks (P includes the links between a-nodes and b-nodes, Q includes the links between b-nodes and c-nodes). Connector nodes belong to b-nodes. An interconnection motif is defined to comprise three sets of connected nodes: the connector nodes (belonging to b-nodes), the nodes in one subnetwork (belonging to a-nodes in the P subnetwork), and the nodes in the other subnetwork (belonging to c-nodes in the Q subnetwork). Each motif has maximumly 6 nodes, resulting in a total of 48 distinct motif forms. The algorithm for counting interconnection motifs is designed by extending the fast approach from Simmons et al.(2019), which uses mathematical operations directly on the bi-adjacency matrix. weighted

If the weighted = FALSE, the input network can be an "igraph" object or two matrices. If a weighted network is provided, it will be transformed to a binary network. If the weighted = TRUE, the input network can only be two matrices. Weighted motifs are considered, then the counts are weighed by the link strength in each motif.

network.or.subnet_mat1 and subnet_mat2

There are two types of inputs network.or.subnet_mat1 that can be processed:

Input is a network of type "igraph" alone.

Must be entered as matrix with subnet_mat2.

If the inputs are two matrices, please make sure the rows of network.or.subnet_mat1 and subnet_mat2 both represent the groups of connector species,i.e, the b-group species. If both matrices have row names, then the function matches row names to produce connector nodes. Otherwise, row numbers are assigned as row names.

References

Pilosof, S., Porter, M. A., Pascual, M., & KC)fi, S. (2017). The multilayer nature of ecological networks. Nature Ecology & Evolution, 1(4), 0101.

Simmons, B. I., Sweering, M. J., Schillinger, M., Dicks, L. V., Sutherland, W. J., & Di Clemente, R. (2019). bmotif: A package for motif analyses of bipartite networks. Methods in Ecology and Evolution, 10(5), 695-701.

#'

Examples

## generate a random tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)
icmotif_count(Net)
#>   M111   M112   M113   M114   M211   M212   M213   M311   M312   M411   M121 
#>     92    132    120     67     71     70     47     47     27     30     22 
#> M122-1 M122-2 M122-3 M123-1 M123-2 M123-3 M123-4 M123-5 M221-1 M221-2 M221-3 
#>    205     97      6    454     99     91     26      0    130     49      5 
#> M222-1 M222-2 M222-3 M222-4 M222-5 M222-6 M222-7 M222-8 M222-9 M321-1 M321-2 
#>    563     28    676     66     20     33      7      1    132    140     47 
#> M321-3 M321-4 M321-5   M131 M132-1 M132-2 M132-3 M132-4 M132-5 M231-1 M231-2 
#>     19     12      0      2     45     10     10      3      0     46      5 
#> M231-3 M231-4 M231-5   M141 
#>      8      2      0      0 

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
icmotif_count(Net)
#>    M111    M112    M113    M114    M211    M212    M213    M311    M312    M411 
#>     612    1341    1754    1836   10801   22680   24570  134437  278640 1269020 
#>    M121  M122-1  M122-2  M122-3  M123-1  M123-2  M123-3  M123-4  M123-5  M221-1 
#>     123     581     335     180    1818     400     516     180     180    2429 
#>  M221-2  M221-3  M222-1  M222-2  M222-3  M222-4  M222-5  M222-6  M222-7  M222-8 
#>    4755     969   25549    3496   10309    6387    2427    2800    6840    1530 
#>  M222-9  M321-1  M321-2  M321-3  M321-4  M321-5    M131  M132-1  M132-2  M132-3 
#>    8124   47962   53347   39968   38250    4930      11     204      55       0 
#>  M132-4  M132-5  M231-1  M231-2  M231-3  M231-4  M231-5    M141 
#>      44       0     219     550      59     154      55       0 

set.seed(12)
MAT <- build_net(11,22,21,0.2,output_matrices=TRUE)
icmotif_count(MAT[["subnet_P"]],MAT[["subnet_Q"]])
#>   M111   M112   M113   M114   M211   M212   M213   M311   M312   M411   M121 
#>    195    368    422    334    159    257    252     91    109     48     75 
#> M122-1 M122-2 M122-3 M123-1 M123-2 M123-3 M123-4 M123-5 M221-1 M221-2 M221-3 
#>    909    448     27   2535    594    647    159      8    520    195     13 
#> M222-1 M222-2 M222-3 M222-4 M222-5 M222-6 M222-7 M222-8 M222-9 M321-1 M321-2 
#>   2518    113   3402    460     81    160     50      7    620    609    168 
#> M321-3 M321-4 M321-5   M131 M132-1 M132-2 M132-3 M132-4 M132-5 M231-1 M231-2 
#>    109     24      1     13    465     90    147     14      0    327     57 
#> M231-3 M231-4 M231-5   M141 
#>     53      9      0      1 

##input as binary matrices,with row names.
md1 <- matrix(sample(c(0,1),8*11,replace=TRUE),8,11,dimnames = list(paste0("b",1:8),paste0("c",1:11)))
md2 <- matrix(sample(c(0,1),10*12,replace=TRUE),10,12,dimnames = list(paste0("b",1:10),paste0("a",1:12)))
icmotif_count(md1,md2)
#> re-check whether the row name of network.or.subnet_mat1 is corresponding to the row name of subnet_mat2!!!
#>   M111   M112   M113   M114   M211   M212   M213   M311   M312   M411   M121 
#>    232    560    838    892    583   1377   1966    857   1936    809    160 
#> M122-1 M122-2 M122-3 M123-1 M123-2 M123-3 M123-4 M123-5 M221-1 M221-2 M221-3 
#>    541    892    143   1300   1204   1100    758     63    479    845    186 
#> M222-1 M222-2 M222-3 M222-4 M222-5 M222-6 M222-7 M222-8 M222-9 M321-1 M321-2 
#>   3246    565   2894   2357   1018    428    730    170   2602   1132    953 
#> M321-3 M321-4 M321-5   M131 M132-1 M132-2 M132-3 M132-4 M132-5 M231-1 M231-2 
#>   1063    876    131     56    312    285    273    163     19    290    241 
#> M231-3 M231-4 M231-5   M141 
#>    226    221     18     11 

##input as weighted matrices,with row numbers as row names.
mdw1 <- matrix(sample(c(rep(0,40),runif(48,0,1))),8,11)
mdw2 <- matrix(sample(c(rep(0,40),runif(80,0,1))),10,12)
icmotif_count(mdw1,mdw2,weighted=T)
#> re-check whether the row name of network.or.subnet_mat1 is corresponding to the row name of subnet_mat2!!!
#>        M111        M112        M113        M114        M211        M212 
#>  90.5588658 144.7964781 136.6950408  82.9787171 133.1731876 199.5948264 
#>        M213        M311        M312        M411        M121      M122-1 
#> 175.3076115 121.1553352 170.3721431  72.6682909  26.4658165  36.2279076 
#>      M122-2      M122-3      M123-1      M123-2      M123-3      M123-4 
#>  56.2389905  13.1459905  33.8054364  22.1862901  29.2731957  23.8907713 
#>      M123-5      M221-1      M221-2      M221-3      M222-1      M222-2 
#>   3.9024768  71.9599042  67.4845325  12.5316711  99.1090961  16.8700423 
#>      M222-3      M222-4      M222-5      M222-6      M222-7      M222-8 
#> 148.2530767  67.6546338  26.5140688  40.8623087  31.4797253  16.1203914 
#>      M222-9      M321-1      M321-2      M321-3      M321-4      M321-5 
#>  80.8262199 104.3050230  42.7922509  37.7682189  28.0544838   3.5063994 
#>        M131      M132-1      M132-2      M132-3      M132-4      M132-5 
#>   3.7953893   5.9869231   3.6005735   4.8063669   4.3622539   0.4701295 
#>      M231-1      M231-2      M231-3      M231-4      M231-5        M141 
#>  18.1568197   7.5448729   8.2151857   3.6088518   0.7342054   0.2074317