Analyzing role of interconnecting node in motifs

Counting the number of 70 roles about interconnecting species in tripartite network motifs.

icmotif_role(network.or.subnet_mat1, subnet_mat2 = NULL, ic_number = 4)

Arguments

network.or.subnet_mat1

Either a tripartite network of 'igraph' class which contains three groups of species and interactions within subnetwork without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species. Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.

subnet_mat2

A numeric matrix(or data.frame) representing interactions between two groups of species. Each row and column of matrix represents single species in the second and third groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise. If network.or.subnet_mat1 is "igraph", subnet_mat2 defaults to NULL.

ic_number

The max number of interconnection code in each motif. Defaults to 4 because of containing all motifs, also can input 2 to calculate the top 36 for rapid computing speed.

Value

If ic_number = 4, return a matrix of 70 columns representing the roles of interconnecting species in the motifs. Columns names are Role1, Role2, Role3 ... Role70.

Wherever, the ic_number = 2, return a matrix of 50 columns representing the roles of interconnecting species in the top 30 motifs. Columns names are Role1, Role2, Role3 ... Role50.

Each row of matrix corresponds to a interconnecting species in the second group of network. If a interconnecting species is linked to both the first and third group species, the elements in this row are not all zero, otherwise the elements are all zero.

Details

network.or.subnet_mat1 and subnet_mat2

There are two types of network.or.subnet_mat1 that can be processed:

Input in a network of type "igraph" alone.

Must be entered as data frame or matrix with subnet_mat2.

If the type of inputting is data frame or matrix, please make sure the row of network.or.subnet_mat1 and subnet_mat2 correspond with the second group of species that both belong to two subnetworks and interact with other groups of species.

Try to make the rows of both matrices have the same attributes. Or we default:

When the two matrices can have different numbers of rows:

1. If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.

2. If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.

When the two matrices can have the same numbers of rows:

No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).

The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!

About a network of type "igraph", It can be obtained from the connection matrices of subnetworks by the function igraph_from_matrices.

References

Simmons, B. I., Sweering, M. J., Schillinger, M., Dicks, L. V., Sutherland, W. J., & Di Clemente, R. (2019). bmotif: A package for motif analyses of bipartite networks. Methods in Ecology and Evolution, 10(5), 695-701.

Examples

## generate a random tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
icmotif_role(Net)
#>                          role1 role2 role3 role4 role5 role6 role7 role8  role9
#> pl_Anthriscus sylvestris     0     0     0     0     0     0     0     0      0
#> pl_Bellis perennis           0     0     0     0     0     0     0     0      0
#> pl_Centaurea nigra           0     0     0     0     0     0     0     0      0
#> pl_Cerastium fontanum        3     0     0     0     3     0     0     1      0
#> pl_Conopodium majus          0     0     0     0     0     0     0     0      0
#> pl_Euphrasia agr             0     0     0     0     0     0     0     0      0
#> pl_Holcus lanatus           14    91   364  1001     0     0     0     0      0
#> pl_Hypochaeris radicata      0     0     0     0     0     0     0     0      0
#> pl_Leontodon autumnalis      0     0     0     0     0     0     0     0      0
#> pl_Leontodon hispidus        0     0     0     0     0     0     0     0      0
#> pl_Leucanthemum vulgare    185   370   370   185  3330  6660  6660 38850  77700
#> pl_Lotus corniculatus        0     0     0     0     0     0     0     0      0
#> pl_Prunella vulgaris         0     0     0     0     0     0     0     0      0
#> pl_Ranunculus acris        230   460   460   230  5175 10350 10350 75900 151800
#> pl_Ranunculus repens       168   420   560   420  2268  5670  7560 19656  49140
#> pl_Rhinanthus minor          0     0     0     0     0     0     0     0      0
#> pl_Taraxacum officinale      0     0     0     0     0     0     0     0      0
#> pl_Trifolium pratense        5     0     0     0    10     0     0    10      0
#> pl_Trifolium repens          6     0     0     0    15     0     0    20      0
#> pl_Veronica chamaedrys       1     0     0     0     0     0     0     0      0
#> pl_Agrostis capillaris       0     0     0     0     0     0     0     0      0
#> pl_Agrostis stolonifera      0     0     0     0     0     0     0     0      0
#> pl_Anthoxanthum odoratum     0     0     0     0     0     0     0     0      0
#> pl_Caltha palustris          0     0     0     0     0     0     0     0      0
#> pl_Cardamine prantesis       0     0     0     0     0     0     0     0      0
#> pl_Festuca rubra             0     0     0     0     0     0     0     0      0
#> pl_Grass                     0     0     0     0     0     0     0     0      0
#> pl_Plantago lanceolata       0     0     0     0     0     0     0     0      0
#> pl_Ranunculus bulbosus       0     0     0     0     0     0     0     0      0
#> pl_Rumex acetosa             0     0     0     0     0     0     0     0      0
#> pl_Taraxacum aggr            0     0     0     0     0     0     0     0      0
#>                          role10 role11 role12 role13 role14 role15 role16
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0     32      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare  330225     29    536    129    116      0    814
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris      815925    106    276    154     64    180    424
#> pl_Ranunculus repens     122850    103    284     52    155    180    580
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         5      4      0      0      0      0      0
#> pl_Trifolium repens          15      4     17      0      0      0      0
#> pl_Veronica chamaedrys        0      0     17      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role17 role18 role19 role20 role21 role22 role23
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum        70      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare     904    226    174    516      0      0      0
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris         384     96     96    256      0    180    180
#> pl_Ranunculus repens        390     78    130    260    180      0    180
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens          35      0      0      0      0      0      0
#> pl_Veronica chamaedrys       35      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role24 role25 role26 role27 role28 role29 role30
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0     32   1102     16
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus            37      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare    1027    648    675    198  12326  11615   3485
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        2030   3000   1236    885   8570   5411   1925
#> pl_Ranunculus repens       1760   1095   2832    843   4536   6282   1566
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         2      4      8      6      0      0      0
#> pl_Trifolium repens           2      8      4      6     85    567      0
#> pl_Veronica chamaedrys        0      0      0      0      0    572      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role31 role32 role33 role34 role35 role36 role37
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus           481    148      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare    4108   4801   2592   2895    870    792      0
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        2520   3920   1920   2244   1245    480   2800
#> pl_Ranunculus repens       3200   1440   1875   1248    312   1155   2800
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens           0      0      0      0      0      0      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role38 role39 role40 role41 role42 role43 role44
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0    666
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare       0      0      0   2904   2700  11106  11655
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        5040   1800   1530   4440   1344  28035  12600
#> pl_Ranunculus repens       1800   5040   1530    780   4080   8820  23040
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      1
#> pl_Trifolium repens           0      0      0      0      0      1      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role45 role46 role47 role48 role49 role50 role51
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare    8325   6948  14760   4770   4392    846     11
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        7410  40980  35280  10170  25020   4640     11
#> pl_Ranunculus repens      37608   5415  29880  23292   8820   4366     11
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         4      0      8     12      6      4      0
#> pl_Trifolium repens           0      4      8      6     12      4      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role52 role53 role54 role55 role56 role57 role58
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0     14      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare       8    176     44     11      0      0      0
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris         200      0      0     55      0      0     44
#> pl_Ranunculus repens        200      0     11     44      0      0     44
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens           0      7      0      0      0      0      0
#> pl_Veronica chamaedrys        0      7      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role59 role60 role61 role62 role63 role64 role65
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare      44      0     86    133    209    341     49
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris           0      0    173     46    253    297     24
#> pl_Ranunculus repens          0      0    179     40     88    462     45
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens           0      0      0      0      0      0      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role66 role67 role68 role69 role70
#> pl_Anthriscus sylvestris      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0
#> pl_Leucanthemum vulgare      10     77     77     55      0
#> pl_Lotus corniculatus         0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0
#> pl_Ranunculus acris          35     22    132     55      0
#> pl_Ranunculus repens         14     55     99     55      0
#> pl_Rhinanthus minor           0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0
#> pl_Trifolium repens           0      0      0      0      0
#> pl_Veronica chamaedrys        0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0
#> pl_Grass                      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0
icmotif_role(Net, ic_number=2)
#>                          role1 role2 role3 role4 role5 role6 role7 role8  role9
#> pl_Anthriscus sylvestris     0     0     0     0     0     0     0     0      0
#> pl_Bellis perennis           0     0     0     0     0     0     0     0      0
#> pl_Centaurea nigra           0     0     0     0     0     0     0     0      0
#> pl_Cerastium fontanum        3     0     0     0     3     0     0     1      0
#> pl_Conopodium majus          0     0     0     0     0     0     0     0      0
#> pl_Euphrasia agr             0     0     0     0     0     0     0     0      0
#> pl_Holcus lanatus           14    91   364  1001     0     0     0     0      0
#> pl_Hypochaeris radicata      0     0     0     0     0     0     0     0      0
#> pl_Leontodon autumnalis      0     0     0     0     0     0     0     0      0
#> pl_Leontodon hispidus        0     0     0     0     0     0     0     0      0
#> pl_Leucanthemum vulgare    185   370   370   185  3330  6660  6660 38850  77700
#> pl_Lotus corniculatus        0     0     0     0     0     0     0     0      0
#> pl_Prunella vulgaris         0     0     0     0     0     0     0     0      0
#> pl_Ranunculus acris        230   460   460   230  5175 10350 10350 75900 151800
#> pl_Ranunculus repens       168   420   560   420  2268  5670  7560 19656  49140
#> pl_Rhinanthus minor          0     0     0     0     0     0     0     0      0
#> pl_Taraxacum officinale      0     0     0     0     0     0     0     0      0
#> pl_Trifolium pratense        5     0     0     0    10     0     0    10      0
#> pl_Trifolium repens          6     0     0     0    15     0     0    20      0
#> pl_Veronica chamaedrys       1     0     0     0     0     0     0     0      0
#> pl_Agrostis capillaris       0     0     0     0     0     0     0     0      0
#> pl_Agrostis stolonifera      0     0     0     0     0     0     0     0      0
#> pl_Anthoxanthum odoratum     0     0     0     0     0     0     0     0      0
#> pl_Caltha palustris          0     0     0     0     0     0     0     0      0
#> pl_Cardamine prantesis       0     0     0     0     0     0     0     0      0
#> pl_Festuca rubra             0     0     0     0     0     0     0     0      0
#> pl_Grass                     0     0     0     0     0     0     0     0      0
#> pl_Plantago lanceolata       0     0     0     0     0     0     0     0      0
#> pl_Ranunculus bulbosus       0     0     0     0     0     0     0     0      0
#> pl_Rumex acetosa             0     0     0     0     0     0     0     0      0
#> pl_Taraxacum aggr            0     0     0     0     0     0     0     0      0
#>                          role10 role11 role12 role13 role14 role15 role16
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0     32      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare  330225     29    536    129    116      0    814
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris      815925    106    276    154     64    180    424
#> pl_Ranunculus repens     122850    103    284     52    155    180    580
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         5      4      0      0      0      0      0
#> pl_Trifolium repens          15      4     17      0      0      0      0
#> pl_Veronica chamaedrys        0      0     17      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role17 role18 role19 role20 role21 role22 role23
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum        70      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare     904    226    174    516      0      0      0
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris         384     96     96    256      0    180    180
#> pl_Ranunculus repens        390     78    130    260    180      0    180
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens          35      0      0      0      0      0      0
#> pl_Veronica chamaedrys       35      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role24 role25 role26 role27 role28 role29 role30
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0     32   1102     16
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus            37      0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare    1027    648    675    198  12326  11615   3485
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        2030   3000   1236    885   8570   5411   1925
#> pl_Ranunculus repens       1760   1095   2832    843   4536   6282   1566
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         2      4      8      6      0      0      0
#> pl_Trifolium repens           2      8      4      6     85    567      0
#> pl_Veronica chamaedrys        0      0      0      0      0    572      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role31 role32 role33 role34 role35 role36 role37
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus           481    148      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare    4108   4801   2592   2895    870    792      0
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        2520   3920   1920   2244   1245    480   2800
#> pl_Ranunculus repens       3200   1440   1875   1248    312   1155   2800
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      0
#> pl_Trifolium repens           0      0      0      0      0      0      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role38 role39 role40 role41 role42 role43 role44
#> pl_Anthriscus sylvestris      0      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0    666
#> pl_Hypochaeris radicata       0      0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0      0
#> pl_Leucanthemum vulgare       0      0      0   2904   2700  11106  11655
#> pl_Lotus corniculatus         0      0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0      0
#> pl_Ranunculus acris        5040   1800   1530   4440   1344  28035  12600
#> pl_Ranunculus repens       1800   5040   1530    780   4080   8820  23040
#> pl_Rhinanthus minor           0      0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0      0
#> pl_Trifolium pratense         0      0      0      0      0      0      1
#> pl_Trifolium repens           0      0      0      0      0      1      0
#> pl_Veronica chamaedrys        0      0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0      0
#>                          role45 role46 role47 role48 role49 role50
#> pl_Anthriscus sylvestris      0      0      0      0      0      0
#> pl_Bellis perennis            0      0      0      0      0      0
#> pl_Centaurea nigra            0      0      0      0      0      0
#> pl_Cerastium fontanum         0      0      0      0      0      0
#> pl_Conopodium majus           0      0      0      0      0      0
#> pl_Euphrasia agr              0      0      0      0      0      0
#> pl_Holcus lanatus             0      0      0      0      0      0
#> pl_Hypochaeris radicata       0      0      0      0      0      0
#> pl_Leontodon autumnalis       0      0      0      0      0      0
#> pl_Leontodon hispidus         0      0      0      0      0      0
#> pl_Leucanthemum vulgare    8325   6948  14760   4770   4392    846
#> pl_Lotus corniculatus         0      0      0      0      0      0
#> pl_Prunella vulgaris          0      0      0      0      0      0
#> pl_Ranunculus acris        7410  40980  35280  10170  25020   4640
#> pl_Ranunculus repens      37608   5415  29880  23292   8820   4366
#> pl_Rhinanthus minor           0      0      0      0      0      0
#> pl_Taraxacum officinale       0      0      0      0      0      0
#> pl_Trifolium pratense         4      0      8     12      6      4
#> pl_Trifolium repens           0      4      8      6     12      4
#> pl_Veronica chamaedrys        0      0      0      0      0      0
#> pl_Agrostis capillaris        0      0      0      0      0      0
#> pl_Agrostis stolonifera       0      0      0      0      0      0
#> pl_Anthoxanthum odoratum      0      0      0      0      0      0
#> pl_Caltha palustris           0      0      0      0      0      0
#> pl_Cardamine prantesis        0      0      0      0      0      0
#> pl_Festuca rubra              0      0      0      0      0      0
#> pl_Grass                      0      0      0      0      0      0
#> pl_Plantago lanceolata        0      0      0      0      0      0
#> pl_Ranunculus bulbosus        0      0      0      0      0      0
#> pl_Rumex acetosa              0      0      0      0      0      0
#> pl_Taraxacum aggr             0      0      0      0      0      0

set.seed(12)
MAT <- build_net(11,22,21,0.2,asmatrices=TRUE)
#> Error in build_net(11, 22, 21, 0.2, asmatrices = TRUE): unused argument (asmatrices = TRUE)

icmotif_role(MAT[[3]],MAT[[4]])
#> Error: object 'MAT' not found

md1 <- matrix(sample(c(0,1),88,replace=TRUE),8,11)
md2 <- matrix(sample(c(0,1),120,replace=TRUE),10,12)
icmotif_role(md1,md2)
#> re-check whether the row name of network.or.subnet_mat1 is corresponding to the row name of subnet_mat2!!!
#>           role1 role2 role3 role4 role5 role6 role7 role8 role9 role10 role11
#> mid_spe1     36    90   120    90    90   225   300   120   300     90     65
#> mid_spe2     18     9     0     0    72    36     0   168    84    252     38
#> mid_spe3     48   120   160   120   168   420   560   336   840    420     79
#> mid_spe4     20    40    40    20    30    60    60    20    40      5     28
#> mid_spe5     30    75   100    75    60   150   200    60   150     30     53
#> mid_spe6     48   120   160   120   168   420   560   336   840    420     94
#> mid_spe7     32   112   224   280    48   168   336    32   112      8     56
#> mid_spe8     25    50    50    25    50   100   100    50   100     25     51
#> mid_spe9      0     0     0     0     0     0     0     0     0      0      0
#> mid_spe10     0     0     0     0     0     0     0     0     0      0      0
#>           role12 role13 role14 role15 role16 role17 role18 role19 role20 role21
#> mid_spe1     155    157    195     65    163    159    165    211    449    179
#> mid_spe2     142    182     28      5      0    300    360      0    142      0
#> mid_spe3     135    141    159    118    185    108     90    140    305    178
#> mid_spe4     178     92     74     19    227    270    130     81    258     30
#> mid_spe5     132    150    143     61    141    114    162    144    378    142
#> mid_spe6     172    170    200    135    208    144    162    190    444    210
#> mid_spe7     126     98    232     80    238     46     42    390    406    306
#> mid_spe8     134    111     70     67    158    179    111     66    220     39
#> mid_spe9       0      0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0      0
#>           role22 role23 role24 role25 role26 role27 role28 role29 role30 role31
#> mid_spe1     168     27    153    155    195     85    377    445    199    415
#> mid_spe2      20      0     34    148     38     78    544    131    296     26
#> mid_spe3     200     98    227    327    189    113    585    305    180    427
#> mid_spe4      55      9    111     44    113     20    202    650    166    260
#> mid_spe5     186     34    178     96    167     58    236    427    146    518
#> mid_spe6     203    110    136    342    160    158    620    244    292    272
#> mid_spe7     140     58    175     66    209     51    128    462    125    703
#> mid_spe8     112     54    168     90    197     57    236    264    150    292
#> mid_spe9       0      0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0      0
#>           role32 role33 role34 role35 role36 role37 role38 role39 role40 role41
#> mid_spe1     403    429    449    191    273    175    173    185     76    403
#> mid_spe2     144    108    171    376     58      4     20      5     10    704
#> mid_spe3     471    639    313    174    237    354    498    286    164    639
#> mid_spe4     401     98    355     64     62     92     39     76      9    148
#> mid_spe5     446    256    389    166    158    186    112    174     66    268
#> mid_spe6     274    700    254    254    350    204    505    244    220    682
#> mid_spe7     290    262    338     91    217    261    100    303     70    112
#> mid_spe8     484    148    371    105     66    190    106    280     81    234
#> mid_spe9       0      0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0      0
#>           role42 role43 role44 role45 role46 role47 role48 role49 role50 role51
#> mid_spe1     605    174    131    217    162    417    279    148     64     45
#> mid_spe2      28     64      9      9    245    144     80    273     91     31
#> mid_spe3     373    424    139    141    573    741    291    408     90     50
#> mid_spe4     300     73    176    204     25    187     76     19      7     12
#> mid_spe5     487    179    290    300     69    354    157     75     33     34
#> mid_spe6     312    204     61     91    510    544    288    516    144     66
#> mid_spe7     857    109    319    385     36    307    160     30     24     32
#> mid_spe8     228    158    260    330     57    356    216     78     31     36
#> mid_spe9       0      0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0      0
#>           role52 role53 role54 role55 role56 role57 role58 role59 role60 role61
#> mid_spe1      97     67     91     89    100     90    102     67     16    112
#> mid_spe2     125     50     16    127    138     18     15     82      0     38
#> mid_spe3     129     27     46    129     66     39    126     27     39     81
#> mid_spe4      87    114     19     22    151     41     29     30      0     47
#> mid_spe5      82     41     55     80    129     65     83     52     16    147
#> mid_spe6     196     52     61    174    103     72    183     37     43    121
#> mid_spe7      53     74     79     41    130    108    103     36     21    160
#> mid_spe8     163     41     15    102     99     25     81     30     24    112
#> mid_spe9       0      0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0      0
#>           role62 role63 role64 role65 role66 role67 role68 role69 role70
#> mid_spe1      36     30    120    148     72     80    131     32     12
#> mid_spe2       9     40     32     36     74     18    152     28     10
#> mid_spe3     129    113     48    239    145     96    177     30     13
#> mid_spe4      35     10     41     71     15     30     16      5      1
#> mid_spe5      48     19    118    175     37     57     83     17      8
#> mid_spe6      37    104    104    137    181     58    264     47     17
#> mid_spe7      37      6    135    184     30     55     68     11      4
#> mid_spe8      78     31    108    198     40     89     75     19     11
#> mid_spe9       0      0      0      0      0      0      0      0      0
#> mid_spe10      0      0      0      0      0      0      0      0      0

R <- rownames(MAT[[4]])[12]
#> Error: object 'MAT' not found
MR <- MAT[[4]][12,]
#> Error: object 'MAT' not found
MAT[[4]] <- MAT[[4]][-12,]
#> Error: object 'MAT' not found
MAT[[4]] <- rbind(MAT[[4]],MR)
#> Error: object 'MAT' not found
rownames(MAT[[4]])[22] <- R
#> Error: object 'R' not found

icmotif_role(MAT[[3]],MAT[[4]])
#> Error: object 'MAT' not found