Transforming two matrices into an igraph object.

Transform two matrices into an igraph object.

igraph_from_matrices(mat1, mat2, weighted = F)

Arguments

mat1

A numeric matrix representing the first subnetwork. Rows should be the shared set of species.

mat2

A numeric matrix representing the second subnetwork. Rows should be the shared set of species.

Value

Return a network of type "igraph".

Details

In this package, a tripartite network contains three groups of nodes (a-nodes,b-nodes,c-nodes) and two subnetworks (P includes the links between a-nodes and b-nodes, Q includes the links between b-nodes and c-nodes). Connector nodes belong to b-nodes. Please make sure the rows of mat1 and mat2 both represent the groups of connector species,i.e, the b-group species. If both matrices have row names, then the function matches row names to define connector nodes. Otherwise, row numbers are assigned to row names and matched, which might produce an incorrected network. Within the two matrices (P and Q), columns represents a-group nodes and c-group nodes respectively. Elements in matrices are non-zero values if two nodes are linked with or without weights, and 0 otherwise.

Examples

set.seed(12)
MAT <- build_net(11,22,21,0.2,output_matrices=T)
MAT[[1]]
#> IGRAPH 0917634 UN-- 54 146 -- 
#> + attr: layout (g/n), name (v/c), level (v/n)
#> + edges from 0917634 (vertex names):
#>  [1] a1 --b1  a1 --b3  a1 --b7  a1 --b19 a1 --b22 a2 --b5  a2 --b8  a2 --b9 
#>  [9] a2 --b13 a2 --b21 a3 --b13 a3 --b17 a3 --b20 a4 --b5  a4 --b9  a4 --b11
#> [17] a4 --b15 a5 --b1  a5 --b5  a5 --b9  a5 --b11 a6 --b1  a6 --b3  a6 --b8 
#> [25] a6 --b21 a7 --b1  a7 --b17 a7 --b18 a8 --b20 a9 --b1  a9 --b2  a9 --b10
#> [33] a9 --b15 a9 --b21 a10--b1  a10--b8  a10--b10 a10--b14 a10--b19 a11--b4 
#> [41] a11--b6  a11--b14 a11--b16 a11--b20 a11--b21 a11--b22 b1 --c8  b1 --c19
#> [49] b1 --c21 b2 --c2  b2 --c3  b2 --c6  b2 --c16 b2 --c17 b2 --c21 b3 --c3 
#> [57] b3 --c5  b3 --c8  b3 --c10 b3 --c17 b4 --c6  b4 --c10 b4 --c14 b4 --c16
#> + ... omitted several edges

tmat<-t(MAT[[3]])
colnames(tmat)<-NULL
igraph_from_matrices(MAT[[3]],MAT[[4]])
#> IGRAPH a90902c UN-- 54 146 -- 
#> + attr: layout (g/n), name (v/c), level (v/n)
#> + edges from a90902c (vertex names):
#>  [1] a1 --b1  a1 --b3  a1 --b7  a1 --b19 a1 --b22 a2 --b5  a2 --b8  a2 --b9 
#>  [9] a2 --b13 a2 --b21 a3 --b13 a3 --b17 a3 --b20 a4 --b5  a4 --b9  a4 --b11
#> [17] a4 --b15 a5 --b1  a5 --b5  a5 --b9  a5 --b11 a6 --b1  a6 --b3  a6 --b8 
#> [25] a6 --b21 a7 --b1  a7 --b17 a7 --b18 a8 --b20 a9 --b1  a9 --b2  a9 --b10
#> [33] a9 --b15 a9 --b21 a10--b1  a10--b8  a10--b10 a10--b14 a10--b19 a11--b4 
#> [41] a11--b6  a11--b14 a11--b16 a11--b20 a11--b21 a11--b22 b1 --c8  b1 --c19
#> [49] b1 --c21 b2 --c2  b2 --c3  b2 --c6  b2 --c16 b2 --c17 b2 --c21 b3 --c3 
#> [57] b3 --c5  b3 --c8  b3 --c10 b3 --c17 b4 --c6  b4 --c10 b4 --c14 b4 --c16
#> + ... omitted several edges
M <- igraph_from_matrices(tmat,MAT[[4]])
#> Error in igraph_from_matrices(tmat, MAT[[4]]): The two networks are not interconnected!
M
#> Error: object 'M' not found