Calculating the number of 48 motifs

Calculating the number of 48 motifs from a tripartite interaction network.

icmotif_count(network.or.subnet_mat1, subnet_mat2 = NULL)

Arguments

network.or.subnet_mat1

Either a tripartite network of 'igraph' class which contains three groups of species and interactions within subnetwork without interactions between each group of species, or a numeric matrix(or data.frame) representing interactions between two groups of species. Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.

subnet_mat2

A numeric matrix(or data.frame) representing interactions between two groups of species. Each row and column of matrix represents single species in the second and third groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise. If network.or.subnet_mat1 is "igraph", subnet_mat2 defaults to NULL.

Value

Return a numeric vector with the number of 48 motifs: M111, M112, M113, M114, M211, M212, M213, M311, M312, M411, M121_1, M122_1, M122_2, M122_3, M123_1, M123_2, M123_3, M123_4, M123_5, M221_1, M221_2, M221_3, M222_1, M222_2, M222_3, M222_4, M222_5, M222_6, M222_7, M222_8, M222_8, M321_1, M321_2, M321_3, M321_4, M321_5, M131, M132-1, M132-2, M132-3, M132-4, M132-5, M231-1, M231-2, M231-3, M231-4, M231-5, M141.

Details

network.or.subnet_mat1 and subnet_mat2

There are two types of network.or.subnet_mat1 that can be processed:

Input in a network of type "igraph" alone.

Must be entered as data frame or matrix with subnet_mat2.

If the type of inputting is data frame or matrix, please make sure the row of network.or.subnet_mat1 and subnet_mat2 correspond with the second group of species that both belong to two subnetworks and interact with other groups of species.

Try to make the rows of both matrices have the same attributes. Or we default:

When the two matrices can have different numbers of rows:

1. If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.

2. If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.

When the two matrices can have the same numbers of rows:

No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).

The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!

About a network of type "igraph", It can be obtained from the connection matrices of subnetworks by the function igraph_from_matrices.

References

Pilosof, S., Porter, M. A., Pascual, M., & Kéfi, S. (2017). The multilayer nature of ecological networks. Nature Ecology & Evolution, 1(4), 0101.

Simmons, B. I., Sweering, M. J., Schillinger, M., Dicks, L. V., Sutherland, W. J., & Di Clemente, R. (2019). bmotif: A package for motif analyses of bipartite networks. Methods in Ecology and Evolution, 10(5), 695-701.

Examples

## generate a random tripartite network
set.seed(12)
Net <- build_net(11,15,16,0.2)

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow
icmotif_count(Net)
#>  [1]     612    1341    1754    1836   10801   22680   24570  134437  278640
#> [10] 1269020     123     581     335     180    1818     400     516     180
#> [19]     180    2429    4755     969   25549    3496   10309    6387    2427
#> [28]    2800    6840    1530    8124   47962   53347   39968   38250    4930
#> [37]      11     204      55       0      44       0     219     550      59
#> [46]     154      55       0

set.seed(12)
MAT <- build_net(11,22,21,0.2,asmatrices=TRUE)
icmotif_count(MAT[[3]],MAT[[4]])
#>  [1]  230  489  649  599  212  471  663   92  223   18  112 1238  745   45 3722
#> [16] 1150 1146  280    6  709  347   20 4005  162 4778 1271  127  287  146    8
#> [31] 1113  963  190  255   46    2   31  860  253  303   53    3  499  112  120
#> [46]   30    2    4

md1 <- matrix(sample(c(0,1),120,replace=TRUE),8,15)
md2 <- matrix(sample(c(0,1),120,replace=TRUE),10,12)
icmotif_count(md1,md2)
#> re-check whether the row name of network.or.subnet_mat1 is corresponding to the row name of subnet_mat2!!!
#>  [1]  320  890 1497 1713  989 2890 5131 1815 5570 2179  251  735 1396  344 1976
#> [16] 1781 1623 1788  267  932 1693  333 4695  964 5107 5245 1903 1201 2385  487
#> [31] 4057 2754 2477 2698 2146  234  118  515  518  467  472   74  671  614  537
#> [46]  540   62   36

R <- rownames(MAT[[4]])[12]
MR <- MAT[[4]][12,]
MAT[[4]] <- MAT[[4]][-12,]
MAT[[4]] <- rbind(MAT[[4]],MR)
rownames(MAT[[4]])[22] <- R

icmotif_count(MAT[[3]],MAT[[4]])
#>  [1]  230  489  649  599  212  471  663   92  223   18  112 1238  745   45 3722
#> [16] 1150 1146  280    6  709  347   20 4005  162 4778 1271  127  287  146    8
#> [31] 1113  963  190  255   46    2   31  860  253  303   53    3  499  112  120
#> [46]   30    2    4