Transforming matrices into network

Two matrices contain three groups of tropical level species. A multilayer network can be transformed from existing matrices data.

igraph_from_matrices(mat1, mat2, isDirected1 = TRUE, isDirected2 = TRUE)

Arguments

mat1

A numeric matrix(or data.frame) representing interactions between two groups of species.Each row and column of matrix represents single species in the second and first groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.

mat2

A numeric matrix(or data.frame) representing interactions between two groups of species.Each row and column of matrix represents single species in the second and third groups of the tripartite network respectively. Elements of matrix are non-zero numbers if the two groups of species are connected, and 0 otherwise.

isDirected1

Logical. Whether the interaction between the two groups of species in mat1 is unidirectional.Default to TRUE, such as Predation and Herbivory. Otherwise it is bidirectional, such as Mutualism.

isDirected2

Logical. Whether the interaction between the two groups of species in mat2 is unidirectional.Default to TRUE, such as Predation and Herbivory. Otherwise it is bidirectional, such as Mutualism.

Value

Return a network of type "igraph".

Details

mat1 and mat2

The type of inputting is data frame or matrix, please make sure the row of mat1 and mat2 correspond with the second group of species that both belong to two subnetworks and interact with other groups of species.

Try to make the rows of both matrices have the same attributes. Or we default:

When the two matrices can have different numbers of rows:

(1). If both matrices have row names, then the function counts all row names to produce two new matrices with the same row names.

(2). If at most one matrix has row names, the function assigns new row names to both matrices on a row-to-row basis (any extra row names are assigned a new value) and then counts all row names to produce two new matrices with the same row names.

When the two matrices can have the same numbers of rows:

No matter how the row names of the two matrices are arranged, as long as the row names are exactly the same; But we don't handle matrices with empty row names (the function will give an error).

The two matrices can have different numbers of rows, but read our default handling carefully to make sure the calculation is accurate when using this function!!!

The columns of mat1 and mat2 could be empty. If empty, the function also defaults to the suggested assignment.

Examples

set.seed(12)
MAT <- build_net(11,22,21,0.2,asmatrices=TRUE)
MAT[[1]]
#> IGRAPH eb46af4 DN-- 54 149 -- 
#> + attr: layout (g/n), name (v/c), level (v/n)
#> + edges from eb46af4 (vertex names):
#>  [1] S2 ->S4  S2 ->S45 S2 ->S40 S2 ->S20 S2 ->S50 S2 ->S25 S26->S14 S26->S20
#>  [9] S26->S19 S16->S18 S16->S39 S16->S40 S16->S29 S16->S37 S16->S41 S16->S25
#> [17] S16->S48 S27->S39 S27->S45 S27->S7  S27->S41 S46->S18 S46->S38 S46->S4 
#> [25] S46->S19 S5 ->S18 S5 ->S38 S5 ->S50 S5 ->S37 S5 ->S41 S5 ->S48 S49->S40
#> [33] S49->S52 S49->S48 S28->S45 S28->S30 S28->S50 S28->S19 S28->S37 S28->S41
#> [41] S28->S52 S34->S7  S34->S29 S8 ->S43 S8 ->S4  S8 ->S40 S51->S43 S51->S13
#> [49] S51->S14 S51->S50 S18->S42 S18->S31 S43->S36 S43->S1  S43->S9  S43->S11
#> [57] S38->S15 S38->S10 S38->S17 S38->S9  S38->S6  S53->S23 S53->S36 S53->S22
#> + ... omitted several edges

tmat<-t(MAT[[3]])
colnames(tmat)<-NULL
igraph_from_matrices(MAT[[3]],MAT[[4]])
#> Because you enter two matrices with the same number of rows, the function will default to them having the same row names. Informed!!
#> IGRAPH 9ab7cf7 DN-- 54 149 -- 
#> + attr: name (v/c), level (v/c)
#> + edges from 9ab7cf7 (vertex names):
#>  [1] S2 ->S4  S2 ->S45 S2 ->S40 S2 ->S20 S2 ->S50 S2 ->S25 S26->S14 S26->S20
#>  [9] S26->S19 S16->S18 S16->S39 S16->S40 S16->S29 S16->S37 S16->S41 S16->S25
#> [17] S16->S48 S27->S39 S27->S45 S27->S7  S27->S41 S46->S18 S46->S38 S46->S4 
#> [25] S46->S19 S5 ->S18 S5 ->S38 S5 ->S50 S5 ->S37 S5 ->S41 S5 ->S48 S49->S40
#> [33] S49->S52 S49->S48 S28->S45 S28->S30 S28->S50 S28->S19 S28->S37 S28->S41
#> [41] S28->S52 S34->S7  S34->S29 S8 ->S43 S8 ->S4  S8 ->S40 S51->S43 S51->S13
#> [49] S51->S14 S51->S50 S18->S42 S18->S31 S43->S36 S43->S1  S43->S9  S43->S11
#> [57] S38->S15 S38->S10 S38->S17 S38->S9  S38->S6  S53->S23 S53->S36 S53->S22
#> + ... omitted several edges
M <- igraph_from_matrices(tmat,MAT[[4]])
#> You will obtain a network that contains non-connecting node!!!
M
#> IGRAPH b39d6ce DN-- 76 149 -- 
#> + attr: name (v/c), level (v/c)
#> + edges from b39d6ce (vertex names):
#>  [1] up_spe1 ->S16 up_spe1 ->S46 up_spe1 ->S5  up_spe2 ->S8  up_spe2 ->S51
#>  [6] up_spe3 ->S46 up_spe3 ->S5  up_spe5 ->S16 up_spe5 ->S27 up_spe6 ->S51
#> [11] up_spe7 ->S2  up_spe7 ->S46 up_spe7 ->S8  up_spe9 ->S2  up_spe9 ->S27
#> [16] up_spe9 ->S28 up_spe10->S28 up_spe11->S26 up_spe11->S51 up_spe12->S27
#> [21] up_spe12->S34 up_spe13->S2  up_spe13->S16 up_spe13->S49 up_spe13->S8 
#> [26] up_spe14->S2  up_spe14->S26 up_spe15->S2  up_spe15->S5  up_spe15->S28
#> [31] up_spe15->S51 up_spe16->S26 up_spe16->S46 up_spe16->S28 up_spe17->S16
#> [36] up_spe17->S34 up_spe18->S16 up_spe18->S5  up_spe18->S28 up_spe19->S16
#> + ... omitted several edges