Null model of multilayer network

The null model could be generated according to different matrix scrambling algorithms for interconnection patterns in the multilayer network

null_model(
  network,
  number = NULL,
  null_type = c("subnetwork1", "subnetwork2", "all", "Savue")
)

Arguments

network

A multilayer(tripartite) network of 'igraph' class. The network contains three groups of species and interactions within layers without interactions between each group of species.

number

A numeric value. The number of null model. Default to NULL representing number 1.

null_type

Logical. Four matrix scrambling algorithms. If null_type = NULL, default to "all".

Value

Return a list contains one or more elements. Each element represent a null model of multilayer network.

Details

null_type

For each of the four types of null models, there are corresponding algorithms. The first type, “subnetwork1”, involved scrambling the adjacency matrix of the first and second groups of the multilayer network.

The second type, “subnetwork2”, focused on scrambling the adjacency matrix of the second and third groups.

Comprehensively, the third type, “all”, blended the approaches of the first two to disarrange the entire network's adjacency matrix, achieving a thorough perturbation of the network's structure.

The last type named “Savue” that disarranged inherent structure in terms of the groups of species connected by each interconnecting species of every subnetworks, thus exhibiting different interconnection patterns.

network

About a network of type "igraph", It can be obtained from the connection matrices of subnetworks by the function igraph_from_matrices

References

Vázquez, D. P., C. J. Melian, N. M. Williams, N. Blüthgen, B. R. Krasnov, and R. Poulin. 2007. Species abundance and asymmetric interaction strength in ecological networks. Oikos 116: 1120-1127.

Sauve, A. M., Thébault, E., Pocock, M. J., & Fontaine, C. (2016). How plants connect pollination and herbivory networks and their contribution to community stability. Ecology, 97(4), 908-917.

Examples

## generate a random multilayer network
set.seed(12)
Net <- build_net(11,15,16,0.2)

data(PPH_Coltparkmeadow)
Net <- PPH_Coltparkmeadow

set.seed(123)
null_model(Net)
#> [[1]]
#> IGRAPH e1a4814 DN-- 156 456 -- 
#> + attr: name (v/c), level (v/n)
#> + edges from e1a4814 (vertex names):
#>  [1] po_Mallochohelea sp.    ->pl_Leucanthemum vulgare
#>  [2] pl_Anthriscus sylvestris->po_Mallochohelea sp.   
#>  [3] pl_Anthriscus sylvestris->po_Opomyza petrei      
#>  [4] po_Opomyza petrei       ->pl_Leucanthemum vulgare
#>  [5] po_Opomyza petrei       ->pl_Prunella vulgaris   
#>  [6] po_Opomyza petrei       ->pl_Ranunculus acris    
#>  [7] po_Opomyza petrei       ->pl_Trifolium repens    
#>  [8] po_Pegoplata aestiva    ->pl_Conopodium majus    
#> + ... omitted several edges
#> 
set.seed(123)
null_model(Net,null_type="subnetwork1")
#> [[1]]
#> IGRAPH 8c6805a DN-- 156 456 -- 
#> + attr: name (v/c), level (v/n)
#> + edges from 8c6805a (vertex names):
#>  [1] po_Mallochohelea sp.    ->pl_Leucanthemum vulgare
#>  [2] pl_Anthriscus sylvestris->po_Mallochohelea sp.   
#>  [3] pl_Anthriscus sylvestris->po_Opomyza petrei      
#>  [4] po_Opomyza petrei       ->pl_Leucanthemum vulgare
#>  [5] po_Opomyza petrei       ->pl_Prunella vulgaris   
#>  [6] po_Opomyza petrei       ->pl_Ranunculus acris    
#>  [7] po_Opomyza petrei       ->pl_Trifolium repens    
#>  [8] po_Pegoplata aestiva    ->pl_Conopodium majus    
#> + ... omitted several edges
#> 
set.seed(123)
null_model(Net,null_type="Savue")
#> [[1]]
#> IGRAPH 6aad910 DN-- 156 456 -- 
#> + attr: name (v/c), level (v/n)
#> + edges from 6aad910 (vertex names):
#>  [1] po_Mallochohelea sp.    ->pl_Leontodon hispidus  
#>  [2] po_Mallochohelea sp.    ->pl_Ranunculus repens   
#>  [3] po_Mallochohelea sp.    ->pl_Taraxacum officinale
#>  [4] pl_Anthriscus sylvestris->po_Mallochohelea sp.   
#>  [5] pl_Anthriscus sylvestris->po_Opomyza petrei      
#>  [6] po_Opomyza petrei       ->pl_Centaurea nigra     
#>  [7] po_Opomyza petrei       ->pl_Cerastium fontanum  
#>  [8] po_Opomyza petrei       ->pl_Lotus corniculatus  
#> + ... omitted several edges
#> 
set.seed(123)
null_model(Net,number=2,null_type="Savue")
#> [[1]]
#> IGRAPH 06eb0e3 DN-- 156 456 -- 
#> + attr: name (v/c), level (v/n)
#> + edges from 06eb0e3 (vertex names):
#>  [1] po_Mallochohelea sp.    ->pl_Leontodon hispidus  
#>  [2] po_Mallochohelea sp.    ->pl_Ranunculus repens   
#>  [3] po_Mallochohelea sp.    ->pl_Taraxacum officinale
#>  [4] pl_Anthriscus sylvestris->po_Mallochohelea sp.   
#>  [5] pl_Anthriscus sylvestris->po_Opomyza petrei      
#>  [6] po_Opomyza petrei       ->pl_Centaurea nigra     
#>  [7] po_Opomyza petrei       ->pl_Cerastium fontanum  
#>  [8] po_Opomyza petrei       ->pl_Lotus corniculatus  
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH fedd211 DN-- 156 456 -- 
#> + attr: name (v/c), level (v/n)
#> + edges from fedd211 (vertex names):
#>  [1] po_Mallochohelea sp.    ->pl_Leontodon autumnalis 
#>  [2] po_Mallochohelea sp.    ->pl_Leontodon hispidus   
#>  [3] po_Mallochohelea sp.    ->pl_Rhinanthus minor     
#>  [4] pl_Anthriscus sylvestris->po_Mallochohelea sp.    
#>  [5] pl_Anthriscus sylvestris->po_Opomyza petrei       
#>  [6] po_Opomyza petrei       ->pl_Anthriscus sylvestris
#>  [7] po_Opomyza petrei       ->pl_Bellis perennis      
#>  [8] po_Opomyza petrei       ->pl_Holcus lanatus       
#> + ... omitted several edges
#>